+++ Extension/src/pages/filtering-log/stores/LogStore.ts
@@ -290,6 +290,9 @@ class LogStore {
     preserveLogEnabled = false;
 
     @observable
+    isPreserveLogModalOpen = false;
+
+    @observable
     selectedEvent: UIFilteringLogEvent | null = null;
 
     @observable
@@ -519,6 +522,13 @@ class LogStore {
         await this.getEventsByTabId(this.selectedTabId);
     };
 
+    @action
+    setIsPreserveLogModalOpen = async (value: boolean) => {
+        runInAction(() => {
+            this.isPreserveLogModalOpen = value;
+        });
+    };
+
     @computed
     get events() {
         const filteredEvents = this.filteringEvents.filter((filteringEvent) => {
@@ -636,6 +646,7 @@ class LogStore {
     @action
     setPreserveLog = async (state: boolean) => {
         await messenger.setPreserveLogState(state);
+
         runInAction(() => {
             this.preserveLogEnabled = state;
         });